test: add test for suricata.dns lua lib
Copy existing tests to a pre8 variant for 7.0 testing.
