+++ tests/dns-lua-rules-pre8/suricata.yaml
@@ -1,0 +1,4 @@
+%YAML 1.1
+---
+
+include: ../../etc/suricata-4.0.3.yaml
