--- tests/lua-output-dns-pre8/README.md
@@ -1,3 +1,0 @@
-Tests the output of DNS being logged by Lua.
-
-PCAPs created by Jason Ish.
