--- tests/lua-output-dns-pre8/suricata.yaml
@@ -1,13 +1,0 @@
-%YAML 1.1
----
-
-include: ../../etc/suricata-3.1.2.yaml
-
-rule-files:
-
-outputs:
-  - lua:
-      enabled: yes
-      scripts-dir: .
-      scripts:
-        - test.lua
