--- tests/lua-output-dns-pre8/test.yaml
@@ -1,19 +1,0 @@
-requires:
-  features:
-    - HAVE_LUA
-  lt-version: 8
-
-pcap: ../lua-output-dns/test.pcap
-
-checks:
-  - shell:
-      args: grep -q "Query TX 0d4f \[\*\*\] block.dropbox.com \[\*\*\] A \[\*\*\] 10.16.1.11:49697 -> 10.16.1.1:53" lua-dns.log
-  - shell:
-      args: cat lua-dns.log | grep Response | grep client-cf.dropbox.com | wc -l
-      expect: 2
-  - shell:
-      args: cat lua-dns.log | grep "Response TX 62b2" | grep NXDOMAIN | wc -l
-      expect: 1
-  - shell:
-      args: grep SOA lua-dns.log | wc -l
-      expect: 1
