+++ tests/lua-output-dns/test.yaml
@@ -1,7 +1,5 @@
 requires:
   min-version: 8
-  features:
-    - HAVE_LUA
 
 checks:
   - shell:
